increase tasks count

+++ mcu_os.c
@@ -11,7 +11,7 @@ static void (*task_p[MCU_OS_TASKS_COUNT])(void);
 volatile static uint16_t task_c[MCU_OS_TASKS_COUNT];
 volatile static uint16_t task_cc[MCU_OS_TASKS_COUNT]={0};
 volatile static uint8_t task_cnt=0;
-volatile static uint8_t GlobalShedulerFlags = 0;
+volatile static uint16_t GlobalShedulerFlags = 0;
 volatile static uint16_t counter = 0;
 
 void mcu_os_loop(void)
